backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
defusedxml==0.7.1
Django==4.2
django-allauth==0.57.0
django-cors-headers==4.0.0
django-filter==23.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
ecdsa==0.19.1
idna==3.11
oauthlib==3.3.1
Pillow==10.0.0
pyasn1==0.6.2
pycparser==2.23
PyJWT==2.11.0
python-dotenv==1.0.0
python-jose==3.5.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==72.1.0 
six==1.17.0
social-auth-app-django==5.2.0
social-auth-core==4.8.3
sqlparse==0.5.5
urllib3==2.6.3